Find files that lie in given scan number range
Relying on a NeXuS filename containing XRF data to have
"xsp3_addetector-xrf_windows-xsp3_addetector1" in it was assuming that a
scan would always be using "addetector1", but the suffix can be from 1
to 5.

Therefore, finding the relevant NeXuS files solely from the range
defined by the start scan number and end scan number, independent of the
detector name, is a more robust approach as it avoids this issue.