Change local repo providing the hyperspy dependency
It is an incredibly hacky thing to continue to do, but for the moment it
works.

Previously, a clone of the hyperspy repo in the I14 area on the DLS
filesystem was being used to provide the necessary hyperspy code for the
mutual informaton alignment method.

However, the method failed to run due to that local hyperspy repo
missing some necessary functions (due to it not being updated for a long
time). A new clone of the hyperspy repo has now been created in the I14
area on the DLS filesystem which is up-to-date, and thus fixes the
errors.