Convert `nxs_dir` to the `Path` type
The `python/3.7` module used at DLS to run the `combine_tomo_data.py`
script in "production" has click v7.0, whereas the development conda
environment being used to test the script had click v8.0.

In the docs for v8.0
https://click.palletsprojects.com/en/8.0.x/parameters/?highlight=click%20path#parameter-types
it is explained that the automatic conversion of options to the `Path`
type that are given `type=click.Path()` is new in v8.0, and thus in v7.0
the conversion from `bytes` or `str` to `Path` doesn't occur. This
results in an error when the `nxs_dir` is expected to be a `Path` object
and the `.exists()` method is attempted to be called.

To be compatible with click v7.0, the `nxs_dir` variable is converted to
a `Path` before executing the rest of the code.