Add filepaths of original data as dataset attribute
This is mainly for debugging purposes. Now that scans can be skipped, in
some situations it can be confusing trying to find the original data
corresponding to a particular image in the stack of projections based
purely on its index in the combined stack.

With the added metadata, given the index of an image in the combined
stack, that same index can be used to look up the original filepath it
came from by using that index in the list of strings stored in the
`origin` attribute of the dataset.