Update import of MI alignment function
The package that the MI alignment function comes from has been installed
into an environment that is available on DLS machines via
`module load python/epsic3.10`.

The import of the function has been updated to reflect the assumption
that this module will be used to make the function available, and thus
the script doesn't need the `sys.path.insert()` to make that function
available anymore.